Using LSTM in tracking unit in Models 1/2
Merged the scan functions for the two models. The
flag tracking_lstm_hidden_dim defines the hidden dimension
of the LSTM. If it is <=0, a linear layer is used in the
tracking unit, otherwise a layer composed of an lstm unit
and a linear layer is used. Currently, the initial state of
the LSTM is initialized to be all zeros.